Nav menu is now persistent component.
Menu + button is rendered inside of App on desktop, and inside the episode
player on mobile. This provides a more consistent experience, and translates
nicely to mobile.

Redux is now handling all state transitions, including if the app is in mobile
mode or not, and whether or not the menu is to be displayed.